backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
python-multipart==0.0.9
passlib[argon2]==1.7.4
pyjwt[crypto]==2.9.0
redis==5.0.8
celery==5.4.0
boto3==1.34.159
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
orjson==3.10.7
tenacity==9.0.0
numpy==2.1.1
scikit-learn==1.5.2
nltk==3.9.1
spacy==3.7.5
PyMuPDF==1.24.9
pytesseract==0.3.13
pdfminer.six==20240706
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.4.0
pdf2image==1.17.0
easyocr==1.7.1
opencv-python-headless==4.10.0.84
xlrd==2.0.1
scipy==1.13.1
gunicorn==21.2.0
